requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
replicate==0.18.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.4.2
aiohttp==3.8.6
requests==2.31.0
aiofiles==23.2.1 